Add deletion of service-info

+++ Jenkinsfile
@@ -28,7 +28,10 @@ pipeline {
             steps {
                 withCredentials([string(credentialsId: 'SERVICE_ID', variable: 'SERVICE_ID'), string(credentialsId: 'API_KEY', variable: 'API_KEY')]) {
                 sh '''
-                    kubectl delete secret service-info
+                    exist_service_info=`kubectl get secret service-info | grep service-info | wc -l`
+                    if [$exist_service_info -eq 0]; then
+                      kubectl delete secret service-info
+                    fi
                     echo -n ${SERVICE_ID} > ./SERVICE_ID
                     echo -n ${API_KEY} > ./API_KEY
                     kubectl create secret generic service-info --from-file=./SERVICE_ID --from-file=./API_KEY
